+++ examples/subscribe.py
@@ -1,9 +1,9 @@
-from pytoncenter import AsyncTonCenterClient, Tx
+from pytoncenter.api import AsyncTonCenterClient
+from pytoncenter.types import Tx
 from pytoncenter.utils import get_opcode, decode_base64
 from tonpy import CellSlice
 import asyncio
 from typing import Dict, Callable, Any
-from datetime import datetime
 
 
 async def handle_jetton_internal_transfer(body: CellSlice):
