+++ examples/transaction_flows.py
@@ -1,4 +1,5 @@
-from pytoncenter import AsyncTonCenterClient, Address
+from pytoncenter.api import AsyncTonCenterClient
+from pytoncenter.address import Address
 import asyncio
 import json
 from pytoncenter.debug import pretty_print_trace_tx, create_named_mapping_func
