+++ tests/parallel.py
@@ -1,5 +1,5 @@
 import pytest
-from pytoncenter import AsyncTonCenterClient
+from pytoncenter.api import AsyncTonCenterClient
 from tonpy import begin_cell
 import asyncio
